pyproject.toml
[tool.poetry]
name = "django-rest-homework"
version = "0.1.0"
description = ""
authors = ["karat <karateev.92@gmail.com >"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.4"
djangorestframework = "^3.16.0"
pillow = "^11.3.0"
python-dotenv = "^1.1.1"
psycopg2-binary = "^2.9.10"
django-filter = "^25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
